--- new-web/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "with-chakra-ui",
-  "version": "0.1.0",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "@chakra-ui/icons": "^1.0.0",
-    "@chakra-ui/react": "^1.0.0",
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "axios-hooks": "^2.3.0",
-    "framer-motion": "^2.9.4",
-    "next": "latest",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
-    "recharts": "^2.0.6",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@types/node": "^14.14.22",
-    "typescript": "^4.1.3"
-  }
-}
