--- web/public/manifest.json
@@ -1,7 +1,0 @@
-{
-  "name": "TrenDuk",
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
