--- web/src/app.css
@@ -1,96 +1,0 @@
-.outline-container {
-  display: grid;
-  grid-template-columns: 7% auto 1px auto 7%;
-  grid-template-rows: 50px auto;
-  grid-template-areas:
-    "header header header header header"
-    ". board divider stat .";
-
-  height: 100%;
-  margin: 0;
-}
-
-/* HEADER css */
-header {
-  grid-area: header;
-  box-shadow: 0 2px 8px #f0f1f2;
-}
-
-.inner {
-  margin: auto 7%;
-
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  position: relative;
-}
-
-.title {
-  font-family: Source Sans Pro, Hiragino Sans GB, Arial, sans-serif;
-  font-size: 2em;
-  font-weight: 200;
-  height: 100%;
-
-  -webkit-user-select: none; /* Safari */
-  -moz-user-select: none; /* Firefox */
-  -ms-user-select: none; /* IE10+/Edge */
-  user-select: none; /* Standard */
-
-  pointer-events: none;
-}
-
-.links a {
-  font-family: Source Sans Pro, Hiragino Sans GB, Arial, sans-serif;
-  font-size: 1.2em;
-  height: 100%;
-  width: 40px;
-
-  text-decoration: none;
-
-  -webkit-user-select: none; /* Safari */
-  -moz-user-select: none; /* Firefox */
-  -ms-user-select: none; /* IE10+/Edge */
-  user-select: none; /* Standard */
-}
-
-.about {
-  height: 100%;
-  width: 10%;
-  float: right;
-  padding-top: 10px;
-}
-
-.divider {
-  grid-area: divider;
-  margin: 15px 0;
-  border-left: 1px solid #dadce0;
-}
-
-.board {
-  margin: 15px 15px 15px 0;
-  grid-area: board;
-  min-width: 530px;
-}
-
-.stat {
-  margin: 15px 0 15px 15px;
-  grid-area: stat;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-betweeen;
-}
-
-.stat div {
-  flex: 1;
-  width: 100%;
-}
-
-html,
-body,
-#root {
-  height: 100%;
-  margin: 0;
-}
