--- web/src/cache/gibo/actions.ts
@@ -1,7 +1,0 @@
-import { IGibo } from "../../types";
-import { GET_GIBOS } from "./types";
-
-export const getGibos = (gibos: Array<IGibo>) => ({
-  type: GET_GIBOS,
-  payload: gibos,
-});
