--- web/src/index.tsx
@@ -1,23 +1,0 @@
-// library imports
-import React from "react";
-import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-
-// local imports
-import App from "./App";
-import { store } from "./cache";
-import reportWebVitals from "./reportWebVitals";
-
-ReactDOM.render(
-  <Provider store={store}>
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  </Provider>,
-  document.getElementById("root")
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
