+++ web/src/store/index.ts
@@ -4,16 +4,14 @@ import thunkMiddleware from "redux-thunk";
 import { composeWithDevTools } from "redux-devtools-extension";
 
 import { nodeReducer } from "./node/reducers";
-import { giboReducer } from "./gibo/reducers";
-import { stateReducer } from "./state/reducers";
+import { currentReducer } from "./current/reducers";
 
 const middlewares = [thunkMiddleware];
 const middlewareEnhancer = applyMiddleware(...middlewares);
 
 const rootReducer = combineReducers({
   node: nodeReducer,
-  gibo: giboReducer,
-  state: stateReducer,
+  current: currentReducer,
 });
 
 export const store = createStore(
